refactor(astro,ml,jupyter,viz): make use of lazy accessor to improve import time of vaex
