+++ packages/vaex-astro/setup.py
@@ -14,15 +14,19 @@ version     = version.__version__
 url         = 'https://www.github.com/maartenbreddels/vaex'
 install_requires_astro = ['vaex-core>=0.1']
 
-setup(name=name + '-astro',
-      version=version,
-      description='Astronomy related transformations and FITS file support',
-      url=url,
-      author=author,
-      author_email=author_email,
-      install_requires=install_requires_astro,
-      license=license,
-      packages=['vaex.astro'],
-      zip_safe=False,
-      entry_points={'vaex.plugin': ['astro = vaex.astro.transformations:add_plugin']}
-      )
+setup(
+    name=name + '-astro',
+    version=version,
+    description='Astronomy related transformations and FITS file support',
+    url=url,
+    author=author,
+    author_email=author_email,
+    install_requires=install_requires_astro,
+    license=license,
+    packages=['vaex.astro'],
+    zip_safe=False,
+    entry_points={
+        'vaex.plugin': ['astro = vaex.astro.legacy:add_plugin'],
+        'vaex.dataframe.accessor': ['astro = vaex.astro.transformations:DataFrameAccessorAstro'],
+    },
+)
