+++ packages/vaex-ml/setup.py
@@ -25,4 +25,5 @@ setup(name=name + '-ml',
       packages=['vaex.ml', 'vaex.ml.incubator', 'vaex.ml.datasets'],
       include_package_data=True,
       zip_safe=False,
-    )
+      entry_points={'vaex.dataframe.accessor': ['ml = vaex.ml:DataFrameAccessorML']}
+)
